Moved state control on form inputs to the search-form component. submitSearch function gets passed in the keywords, subreddits, and toggleInbox values as arguments from the search form component.
